requirements.txt
flask
pyodbc
azure-storage-blob
python-dotenv
os
